requirements.txt
mysql_connector_repackaged==0.3.1
Pillow==10.4.0
python_bcrypt==0.3.2
qrcode==7.4.2
streamlit==1.38.0
streamlit==1.33.0
